♻️ F01a: HOME > Navbar > create <Icon/> for reusable-ui and navbar folder
