+++ src/components/pages/home/Header.js
@@ -1,5 +1,5 @@
 import styled from 'styled-components';
-import Navbar from './Navbar';
+import Navbar from './navbar/Navbar';
 
 export default function Header() {
   return (
